添加使用Redis控制用户登录状态有效时长 添加退出登录接口

+++ app/user/cmd/api/internal/config/config.go
@@ -12,7 +12,7 @@ type Config struct {
 	// JWT
 	Auth struct {
 		AccessSecret string
-		AccessExpire int64
+		AccessExpire int
 	}
 	// Redis
 	RedisConf redis.RedisConf
